+++ BloonsTD6 Mod Helper/Api/Commands/Export/ExportGameDataCommand.cs
@@ -2,7 +2,7 @@ using BTD_Mod_Helper.Api.Helpers;
 using CommandLine;
 namespace BTD_Mod_Helper.Api.Commands.Export;
 
-internal class ExportGameModelCommand : ModCommand<ExportCommand>
+internal class ExportGameDataCommand : ModCommand<ExportCommand>
 {
     public override string Command => "gamedata";
     public override string Help => "Exports most static game data to the sandbox root folder";
@@ -10,9 +10,12 @@ internal class ExportGameModelCommand : ModCommand<ExportCommand>
     [Option("clean", Default = false, HelpText = "remove any existing data from the folders")]
     public bool Clean { get; set; } = false;
 
-    [Option("consistent", Default = false,
-        HelpText = "changes the output slightly so that it will be consistent if reserialized")]
-    public bool Consistent { get; set; } = false;
+    [Option("consistent", HelpText = "changes the output slightly so that it will be consistent if reserialized"
+#if DEBUG
+        , Default = true
+#endif
+    )]
+    public bool Consistent { get; set; }
 
     public override bool Execute(ref string resultText)
     {