+++ BloonsTD6 Mod Helper/Api/Components/MatchLocalPosition.cs
@@ -3,13 +3,13 @@ using UnityEngine;
 namespace BTD_Mod_Helper.Api.Components;
 
 /// <summary>
-/// Component to make this transform continuously match the position of another transform
+/// Component to make this transform continuously match the local position of another transform
 /// </summary>
 [RegisterTypeInIl2Cpp(false)]
 public class MatchLocalPosition : MonoBehaviour
 {
     /// <summary>
-    /// Other transform to constantly copy the scale from
+    /// Other transform to constantly copy the local position from
     /// </summary>
     public Transform transformToCopy;